cleaned up UTCSBouncyTableViewCell
Former-commit-id: 74511165a2f05fb303be37b314592a5cc9b6571e
Former-commit-id: 16577657555a13f04ea6ad7269f3c92240e1855b
Former-commit-id: 437777956e9b6e99d5e443233ff2dc10cd375aa9

+++ UTCS/UTCS/Classes/Model/UTCSAuthenticationManager.h
@@ -13,7 +13,7 @@ typedef void (^UTCSAuthenticationCompletion) (BOOL success, NSError *error);
 
 extern NSString * const UTCSAuthenticationErrorDomain;
 
-typedef NS_ENUM(u_int16_t, UTCSAuthenticationErrorCode) {
+typedef NS_ENUM(int8_t, UTCSAuthenticationErrorCode) {
     UTCSAuthenticationConnectionErrorCode   = -1,
     UTCSAuthenticationAccessDeniedErrorCode = -2
 };
